refactor/enhancements (#98)
* #55 refactor: avoid arror on close window and enhance keyboard-input

* #55 refactor: simplify message-sending
